+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # SDI changelogs
 
-## 0.4.0
+## 0.4.0-rc
 
 **ALL PROJECT HAS MIGRATED FROM PACKAGE `com.insanusmokrassar` TO `dev.inmo`**